[FIX] runbot_travis2docker: Skip method if there are not builds
Cleanup original method process cleaning even if you don't have builds.
https://github.com/odoo/runbot/blob/e815571d14163c8af6c19637c244ff022abaa883/runbot/models/build.py#L330-L363

Then we need to skip the original method if there are not builds.

Note: The original method run a query without postgresql port, if you
have changed it then will raise an error.

Uses -v in order to remove the volumes of the container